requirements.txt
bs4==0.0.1
feedparser==6.0.10
html5lib==1.1
requests==2.28.1
urllib3==1.26.5
python-dateutil==2.8.2